+++ src/app.module.ts
@@ -5,7 +5,7 @@ import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { LoginController } from './controllers/login.controller';
-import { LoginService } from './login.service';
+import { LoginService } from './factories/login.service';
 
 @Module({
   imports: [
